+++ CHANGES.rst
@@ -4,4 +4,11 @@ Changelog
 1.0a1 (unreleased)
 -------------------
 
-- Initial release
+- Fix unicode encoding in download csv
+  [gborelli]
+
+- Added Italian translation
+  [gborelli]
+
+- Added fields parameter in Mailer override expressions
+  [gborelli]